refactor: centralize base class imports
