+++ src/dpf2/core/__init__.py
@@ -1,13 +1,13 @@
-"""Core components for simplified DPF simulations."""
+"""Core components for simplified DPF simulations.
+
+Base classes for solver components now live in :mod:`dpf2.core.bases` and
+are no longer re-exported here.
+"""
 
 from .config import DPFConfig
 from .simulation import DPFSimulation
-from .bases import PlasmaSolverBase, CircuitSolverBase, DiagnosticsBase
 
 __all__ = [
     "DPFConfig",
     "DPFSimulation",
-    "PlasmaSolverBase",
-    "CircuitSolverBase",
-    "DiagnosticsBase",
 ]
